+++ src/nvector/_acknowledgements.py
@@ -1,0 +1,14 @@
+"""
+The `nvector package <http://pypi.python.org/pypi/nvector/>`_ for
+`Python <https://www.python.org/>`_ was written by Per A. Brodtkorb at
+`FFI (The Norwegian Defence Research Establishment) <http://www.ffi.no/en>`_
+based on the `nvector toolbox <http://www.navlab.net/nvector/#download>`_ for
+`Matlab <http://www.mathworks.com>`_ written by the navigation group at
+`FFI <http://www.ffi.no/en>`_.
+
+Most of the content is based on the article by K. Gade :cite:`Gade2010Nonsingular`:
+
+Thus this article should be cited in publications using this page or the
+downloaded program code.
+
+"""