+++ src/nvector/_documentation.py
@@ -1,0 +1,21 @@
+"""
+Documentation and code
+======================
+
+Official documentation:
+
+http://www.navlab.net/nvector/
+
+http://nvector.readthedocs.io/en/latest/
+
+*Kenneth Gade (2010):*
+    `A Nonsingular Horizontal Position Representation,
+    The Journal of Navigation, Volume 63, Issue 03, pp 395-417, July 2010.
+    <http://www.navlab.net/Publications/A_Nonsingular_Horizontal_Position_Representation.pdf>`_
+
+
+Bleeding edge: https://github.com/pbrod/nvector.
+
+Official releases available at: http://pypi.python.org/pypi/nvector.
+
+"""