+++ src/nvector/_images.py
@@ -1,0 +1,20 @@
+"""
+.. only:: html
+
+    .. |pkg_img| image:: https://badge.fury.io/py/nvector.png
+       :target: https://pypi.python.org/pypi/nvector/
+    .. |tests_img| image:: https://travis-ci.org/pbrod/Nvector.svg?branch=master
+       :target: https://travis-ci.org/pbrod/Nvector
+    .. |docs_img| image:: https://readthedocs.org/projects/pip/badge/?version=stable
+       :target: http://Nvector.readthedocs.org/en/stable/
+    .. |health_img| image:: https://api.codeclimate.com/v1/badges/c04214bef610b25906fe/maintainability
+       :target: https://codeclimate.com/github/pbrod/Nvector/maintainability
+       :alt: Maintainability
+    .. |coverage_img| image:: https://codecov.io/gh/pbrod/nvector/branch/master/graph/badge.svg
+       :target: https://codecov.io/gh/pbrod/nvector
+    .. |versions_img| image:: https://img.shields.io/pypi/pyversions/Nvector.svg
+       :target: https://github.com/pbrod/nvector
+    .. |downloads_img| image:: https://img.shields.io/pypi/dm/nvector.svg
+       :alt: PyPI - Downloads
+
+"""